Fix cross-compilation
Dependencies were specified in `depsTargetTarget` which would make them end up in the
dependencies for the target platform when they are actually dependencies needed on the
buildPlatform using the rust package. I.e. they are target dependencies of the compiler
package but will be used as build platform dependencies when propagated.